requirements.txt
git+https://github.com/nkgilley/fast.com.git
python-dateutil
py-zabbix

